pyproject.toml
[tool.poetry]
name = "plassembler"
version = "1.5.0" # change VERSION too
description = "Quickly and accurately assemble plasmids in hybrid sequenced bacterial isolates"
authors = ["George Bouras <george.bouras@adelaide.edu.au>"]
license = "MIT"
readme = "README.md"
homepage = "https://github.com/gbouras13/plassembler"
repository = "https://github.com/gbouras13/plassembler"
keywords = ["microbial", "bioinformatics", "plasmids", "bacteria"]
include = [
    "HISTORY.md"
]

[tool.poetry.scripts]
plassembler = 'plassembler:main'

[tool.poetry.dependencies]
python = ">=3.8,<3.10"
click = ">=8.0.0"
loguru = ">=0.5.3"
pyyaml = ">=6.0"
pandas = ">=1.4.2"
biopython = ">=1.76"
pysam = ">=0.16.0"
alive-progress = ">=3.0.1"
requests = ">=2.25.1"


[tool.poetry.dev-dependencies]
black = ">=22.3.0"
isort = ">=5.10.1"
pytest = ">=6.2.5"
pytest-cov = ">=3.0.0"
flake8 = ">=3.0.1"

[[tool.poetry.source]]
name = "pypi-test"
url = "https://test.pypi.org/simple/"
priority = "primary"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
profile = "black"
